splitted Neural Network to own module + movedDatasetSplitter to tools
